update-day4

+++ Lession4/modules/brand/services/brand-service.ts
@@ -11,27 +11,35 @@ export class BrandService implements IBrandService {
         this.brandRepo = brandRepo;
     }
 
-    async createBrandAsync(value: BrandCreateDTO): Promise<void> {
+    async createBrandAsync(value: BrandCreateDTO): Promise<string> {
         await this.brandRepo.addAsync(new Brand(value.name));
+        return "Tạo hãng thành công";
     }
 
-    async updateBrandAsync(value: BrandUpdateDTO): Promise<void> {
+    async updateBrandAsync(value: BrandUpdateDTO): Promise<string> {
+        const data = await this.brandRepo.getByIdAsync(value.id);
+        if (!data) return Promise.reject(Error('Hãng không tồn tại'));
         await this.brandRepo.editAsync(value.id, { name: value.name });
+        return "Cập nhật hãng thành công";
     }
 
-    async getBrand(search: string, pageIndex: number = 1, pageSize: number = 10): Promise<PaginationResult<BrandViewDTO>> {
+    async getBrand(search?: string, pageIndex?: number, pageSize?: number): Promise<PaginationResult<BrandViewDTO>> {
         const brands = await this.brandRepo.getAsync(pageIndex, pageSize, (b) => {
-            return b.ilike('name', `%${search}%`);
+            return b.select('*').ilike('name', `%${search}%`);
         });
         return Pagination<BrandViewDTO>(brands.data.map((b: Brand) => toBrandViewDTO(b)), brands.count, pageSize, pageIndex);
     }
 
     async getBrandById(id: string): Promise<BrandViewDTO> {
         const data = await this.brandRepo.getByIdAsync(id);
+        if (!data) return Promise.reject(Error('Hãng không tồn tại'));
         return toBrandViewDTO(data);
     }
 
-    async deleteBrand(id: string): Promise<void> {
+    async deleteBrand(id: string): Promise<string> {
+        const data = await this.brandRepo.getByIdAsync(id);
+        if (!data) return Promise.reject(Error('Hãng không tồn tại'));
         await this.brandRepo.removeAsync(id);
+        return "Xóa hãng thành công";
     }
 }
